Tax Problem Solved
There are some variations in calculations though we are getting the exact CPP and EI amount I am not getting the proper formula for finding federal tax and other stuff so , finally I am getting the nearest amount right now. It can be solved by the help of officials and as this is prototype we can't deal with that!