Update 2.6.1
Projects will be listed to owner and he/she can assign the project to Developers Team.
This action will create project in Project Hub so details will be shown there and the owner will get option to give priority based on Client Name and Price. Useful if references are given and if owner decide to give high priority to highest valued project